add abstraction layer for model selection in preparation for Hailo implementation
