+++ src/lib/socket_trigger.py
@@ -1,13 +1,14 @@
 
-import json
-import numpy as np
-import cv2
 import base64
 import datetime
+import json
 
+import cv2
+import numpy as np
 from websockets.sync.client import connect
 
 from src.utils.logging import logging_default
+
 
 class SocketTrigger:
     def __init__(self, api_settings_path : str):
@@ -39,7 +40,7 @@ class SocketTrigger:
         # Construct the WebSocket URL
         self.ws_url = f"ws://{self.server_ip}?vehicle_id={self.vehicle_id}&device={self.device_name}"
 
-    def save_image(self, image : np.ndarray, event, target : str = '', wsEvent : str = '') -> None:
+    def save_image(self, image : np.ndarray, event, target : str = '', ws_event : str = '') -> None:
         """
         Save the image locally, encode it in base64, and send it to the WebSocket server.
 
@@ -54,7 +55,7 @@ class SocketTrigger:
         target : str, optional
             The target associated with the event (default is '').
         
-        wsEvent : str, optional
+        ws_event : str, optional
             The WebSocket event type (default is '').
 
         Raises
@@ -64,7 +65,7 @@ class SocketTrigger:
         """
         try:
             # Log the beginning of the process
-            logging_default.info(f"Saving image for event: {event}, target: {target}, websocket event: {wsEvent}")
+            logging_default.info(f"Saving image for event: {event}, target: {target}, websocket event: {ws_event}")
 
             if self.send_to_server:
                 with connect(self.ws_url) as websocket:
@@ -74,8 +75,8 @@ class SocketTrigger:
 
                     with open(f"image_event/frame{timestamp}.jpg", "rb") as image_file:
                         encoded_string = base64.b64encode(image_file.read())
-                        jsonData = {
-                            "event" : wsEvent,
+                        json_data = {
+                            "event" : ws_event,
                             "vehicle_id" : self.vehicle_id,
                             "target" : target,
                             "data" : {
@@ -84,11 +85,11 @@ class SocketTrigger:
                                 "behavior_type" : event
                             }
                         }
-                        websocket.send(json.dumps(jsonData))
+                        websocket.send(json.dumps(json_data))
                         message = websocket.recv()
                         logging_default.info(f"Image saved successfully for event: {event}. Message : {message}")
             else:
-                logging_default.info(f"Will not send any event to server. It's on DEBUG mode!")
+                logging_default.info("Will not send any event to server. It's on DEBUG mode!")
 
         except Exception as e:
             logging_default.error(f"Error saving image for event: {event}. Error: {e}")
