+++ src/routers/video_stream.py
@@ -1,6 +1,10 @@
 from fastapi import APIRouter
 from fastapi.responses import StreamingResponse
-from src.services.drowsiness_detection_service import generate_drowsiness_stream, generate_original_capture_stream
+
+from src.services.drowsiness_detection_service import (
+    generate_drowsiness_stream,
+    generate_original_capture_stream,
+)
 
 router = APIRouter()
