+++ src/utils/drawing_utils.py
@@ -1,5 +1,6 @@
 import cv2
 
+
 def draw_eye_landmarks(image, eye_landmarks, color=(0, 255, 0)):
     for i in range(len(eye_landmarks) - 1):
         pt1 = eye_landmarks[i]
